+++ bridge/platform.go
@@ -1,127 +1,100 @@
 package bridge
 
 import (
+	"encoding/json"
 	"sync"
 
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 )
 
-// PlatformBridge is the interface that describes a bridge to communicate from
-// Go to an underlying platform.
-type PlatformBridge interface {
-	// Request issues a request to the specified URL with the payload.
-	Request(url string, p Payload) (res Payload, err error)
+// PlatformRPCHandler represents the handler that will perform platform calls.
+type PlatformRPCHandler func(call string) error
 
-	// RequestWithAsyncResponse issues a request to the specified URL with the
-	// payload.
-	// It should be used when the response require to wait platform asynchronous
-	// operations.
-	RequestWithAsyncResponse(url string, p Payload) (res Payload, err error)
-
-	// Return returns the response with the identifier generated by a call to
-	// RequestWithAsyncResponse.
-	// It should be called to return a response requested with
-	// RequestWithAsyncResponse.
-	Return(returnID string, p Payload, err error)
-}
-
-// PlatformHandler describes the func that will handle requests to the
+// PlatformRPC is a struct that implements the remote procedure call from  Go to an
 // underlying platform.
-type PlatformHandler func(url string, p Payload, returnID string) (res Payload, err error)
+type PlatformRPC struct {
+	Handler PlatformRPCHandler
 
-// NewPlatformBridge creates an underlying platform bridge.
-func NewPlatformBridge(h PlatformHandler) PlatformBridge {
-	return newPlatformBridge(h)
+	mutex   sync.RWMutex
+	returns map[string]chan rpcReturn
 }
 
-type platformBridge struct {
-	handler PlatformHandler
-	returns *returnRegistry
-}
+// Call calls the given method with the given input and stores the result in
+// the value pointed by the output.
+// It returns an error if the output is not a pointer.
+func (r *PlatformRPC) Call(method string, out interface{}, in interface{}) error {
+	returnID := uuid.New().String()
 
-func newPlatformBridge(h PlatformHandler) *platformBridge {
-	return &platformBridge{
-		handler: h,
-		returns: newReturnRegistry(),
-	}
-}
-
-func (b *platformBridge) Request(rawurl string, p Payload) (res Payload, err error) {
-	res, err = b.handler(rawurl, p, "")
-	return
-}
-
-func (b *platformBridge) RequestWithAsyncResponse(rawurl string, p Payload) (res Payload, err error) {
-	returnID := uuid.New()
-
-	retchan := make(chan returnPayload, 1)
-	defer close(retchan)
-
-	b.returns.Set(returnID, retchan)
-	defer b.returns.Delete(returnID)
-
-	if _, err = b.handler(rawurl, p, returnID.String()); err != nil {
-		return
+	call, err := json.Marshal(PlatformCall{
+		Method:   method,
+		Input:    in,
+		ReturnID: returnID,
+	})
+	if err != nil {
+		return err
 	}
 
-	ret := <-retchan
-	res = ret.response
-	err = ret.err
-	return
+	rpcRetC := make(chan rpcReturn, 1)
+
+	r.mutex.Lock()
+	if r.returns == nil {
+		r.returns = make(map[string]chan rpcReturn)
+	}
+	r.returns[returnID] = rpcRetC
+	r.mutex.Unlock()
+
+	if err = r.Handler(string(call)); err != nil {
+		return err
+	}
+
+	rpcRet := <-rpcRetC
+
+	r.mutex.Lock()
+	delete(r.returns, returnID)
+	close(rpcRetC)
+	r.mutex.Unlock()
+
+	if rpcRet.Error != nil {
+		return rpcRet.Error
+	}
+
+	if len(rpcRet.Output) != 0 {
+		return json.Unmarshal([]byte(rpcRet.Output), out)
+	}
+	return nil
 }
 
-func (b *platformBridge) Return(returnID string, res Payload, err error) {
-	ret := returnPayload{
-		response: res,
-		err:      err,
-	}
+// Return returns the given output to the call that waits for the given return
+// id.
+func (r *PlatformRPC) Return(retID string, out string, errString string) {
+	r.mutex.RLock()
+	rpcRetC, ok := r.returns[retID]
+	r.mutex.RUnlock()
 
-	var retID uuid.UUID
-	if retID, err = uuid.Parse(returnID); err != nil {
-		panic(errors.Wrapf(err, "returning result %s failed", res.String()))
-	}
-
-	retchan, ok := b.returns.Get(retID)
 	if !ok {
-		panic(errors.Errorf("returning result %s failed: no return set for %v",
-			res.String(),
-			retID))
+		panic("no async call for " + retID)
 	}
-	retchan <- ret
-}
 
-type returnPayload struct {
-	response Payload
-	err      error
-}
+	var err error
+	if len(errString) != 0 {
+		err = errors.New(errString)
+	}
 
-type returnRegistry struct {
-	mutex   sync.Mutex
-	returns map[uuid.UUID]chan returnPayload
-}
-
-func newReturnRegistry() *returnRegistry {
-	return &returnRegistry{
-		returns: make(map[uuid.UUID]chan returnPayload),
+	rpcRetC <- rpcReturn{
+		Output: out,
+		Error:  err,
 	}
 }
 
-func (r *returnRegistry) Set(id uuid.UUID, retchan chan returnPayload) {
-	r.mutex.Lock()
-	r.returns[id] = retchan
-	r.mutex.Unlock()
+// PlatformCall represents a platform call.
+type PlatformCall struct {
+	Method   string
+	Input    interface{} `json:",omitempty"`
+	ReturnID string
 }
 
-func (r *returnRegistry) Get(id uuid.UUID) (retchan chan returnPayload, ok bool) {
-	r.mutex.Lock()
-	retchan, ok = r.returns[id]
-	r.mutex.Unlock()
-	return
-}
-
-func (r *returnRegistry) Delete(id uuid.UUID) {
-	r.mutex.Lock()
-	delete(r.returns, id)
-	r.mutex.Unlock()
+type rpcReturn struct {
+	Output string
+	Error  error
 }
